Pushing Change out
Exercise 1 now displays name/date/avatar/id and a button that will toggle date/id for each item on the list.

Exercise 2 now has js and css on thier own file and it will save the list in localstorage. Pressing the Reset button will clear localstorage and reload the page.